cleaning up - added toasts to review page - changed search bar to component
