+++ frontend/src/components/NavBar.js
@@ -8,9 +8,9 @@ const NavBar = () => {
             style={{
                 height: "100vh",
                 width: "75px",
-                position: "fixed",
+                position: "absolute",
                 left: "0",
-                top: "0",
+                bottom: "0", // added to position NavBar at the bottom of the screen
                 backgroundColor: "black",
                 display: "flex",
                 flexDirection: "column",
@@ -36,6 +36,32 @@ const NavBar = () => {
                     size="lg"
                     mt="10px"
                     mb="10px"
+                />
+            </Link>
+            <div style={{ flexGrow: 1 }}></div> {/* added to push last icon to bottom */}
+            <Link href="/">
+                <IconButton
+                    aria-label="Profile"
+                    icon={
+                        <svg
+                            xmlns="http://www.w3.org/2000/svg"
+                            fill="none"
+                            viewBox="0 0 24 24"
+                            strokeWidth={1.5}
+                            stroke="currentColor"
+                            className="h-[48px] w-[48px]"
+                        >
+                            <path
+                                strokeLinecap="round"
+                                strokeLinejoin="round"
+                                d="M15.75 9V5.25A2.25 2.25 0 0013.5 3h-6a2.25 2.25 0 00-2.25 2.25v13.5A2.25 2.25 0 007.5 21h6a2.25 2.25 0 002.25-2.25V15M12 9l-3 3m0 0l3 3m-3-3h12.75"
+                            />
+                        </svg>
+                    }
+                    colorScheme="white"
+                    size="lg"
+                    mt="10px"
+                    mb="20px"
                 />
             </Link>
         </div>
